Finished 3 and added helper function
Command 3 allows the user to update the scheduler system of some digital display. The user choses which digital display to update by inputting the corresponding serial number and then choosing which schedular system to update to. If an invalid serial number or schedulere system is give, program will keep asking for a new one.

Helper function just checks if a given serial number exist which means that the dig display itself also exist.