Fiished 3 and added method that chesk if model with users model no exist. If not a new method cerates the model
command 3 now adds a digital display to the current list of dd's. Inserted dd doesn't seem to save after you log out tho. It stays in the db until you log and up till that point. You can manipulate the newly created dd in the same ways as the dd's that were alreeady in thee db